#16: move Criterion factory to base class
