--- src/Execution/lbsCriterion.py
@@ -1,109 +1,0 @@
-#
-#@HEADER
-###############################################################################
-#
-#                                lbsCriterion.py
-#                           DARMA Toolkit v. 1.0.0
-#               DARMA/LB-analysis-framework => LB Analysis Framework
-#
-# Copyright 2019 National Technology & Engineering Solutions of Sandia, LLC
-# (NTESS). Under the terms of Contract DE-NA0003525 with NTESS, the U.S.
-# Government retains certain rights in this software.
-#
-# Redistribution and use in source and binary forms, with or without
-# modification, are permitted provided that the following conditions are met:
-#
-# * Redistributions of source code must retain the above copyright notice,
-#   this list of conditions and the following disclaimer.
-#
-# * Redistributions in binary form must reproduce the above copyright notice,
-#   this list of conditions and the following disclaimer in the documentation
-#   and/or other materials provided with the distribution.
-#
-# * Neither the name of the copyright holder nor the names of its
-#   contributors may be used to endorse or promote products derived from this
-#   software without specific prior written permission.
-#
-# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
-# AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
-# IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
-# ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE
-# LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
-# CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
-# SUBSTITUTE GOlODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
-# INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
-# CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
-# ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
-# POSSIBILITY OF SUCH DAMAGE.
-#
-# Questions? Contact darma@sandia.gov
-#
-###############################################################################
-#@HEADER
-#
-########################################################################
-lbsCriterion_module_aliases = {}
-for m in [
-    "importlib",
-    ]:
-    has_flag = "has_" + m
-    try:
-        module_object = __import__(m)
-        if m in lbsCriterion_module_aliases:
-            globals()[lbsCriterion_module_aliases[m]] = module_object
-        else:
-            globals()[m] = module_object
-        globals()[has_flag] = True
-    except ImportError as e:
-        print("** ERROR: failed to import {}. {}.".format(m, e))
-        globals()[has_flag] = False
-
-########################################################################
-class Criterion(object):
-    """A helper factory to instantiate desired concrete criteria
-    """
-
-    ####################################################################
-    @staticmethod
-    def factory(criterion_idx, processors, edges, parameters=None):
-        """Produce the necessary concrete criterion
-        """
-
-        # Ensure that criterion index is valid
-        c_name = {
-            0: "GrapevineCriterion",
-            1: "ModifiedGrapevineCriterion",
-            2: "StrictLocalizingCriterion",
-            3: "RelaxedLocalizingCriterion",
-            }.get(criterion_idx)
-        if not c_name:
-            print("** ERROR: unsupported criterion index: {}".format(
-                criterion_idx))
-            return None
-
-        #Try to load corresponding module
-        m_name = "Execution.lbs{}".format(c_name)
-        module = importlib.import_module(m_name)
-        try:
-            module = importlib.import_module(m_name)
-        except:
-            print("** ERROR: could not load module `{}`".format(
-                m_name))
-            return None
-
-        # Try to get concrete criterion class from module
-        try:
-            c_class = getattr(module, c_name)
-        except:
-            print("** ERROR: could not get class `{}` from module `{}`".format(
-                c_name,
-                m_name))
-            return None
-
-        # Instantiate and return object
-        ret_object = c_class(processors, edges, parameters)
-        print("[Criterion] Instantiated {} load transfer criterion".format(
-            c_name))
-        return ret_object
-
-########################################################################
