+++ src/Execution/lbsCriterionBase.py
@@ -44,7 +44,8 @@
 ########################################################################
 lbsCriterionBase_module_aliases = {}
 for m in [
-    "abc"
+    "abc",
+    "importlib",
     ]:
     has_flag = "has_" + m
     try:
@@ -104,6 +105,49 @@ class CriterionBase:
         print("[CriterionBase] Assigned {} processors and {} edges to base criterion".format(
             n_p,
             n_e))
+
+    ####################################################################
+    @staticmethod
+    def factory(criterion_idx, processors, edges, parameters=None):
+        """Produce the necessary concrete criterion
+        """
+
+        # Ensure that criterion index is valid
+        c_name = {
+            0: "GrapevineCriterion",
+            1: "ModifiedGrapevineCriterion",
+            2: "StrictLocalizingCriterion",
+            3: "RelaxedLocalizingCriterion",
+            }.get(criterion_idx)
+        if not c_name:
+            print("** ERROR: unsupported criterion index: {}".format(
+                criterion_idx))
+            return None
+
+        #Try to load corresponding module
+        m_name = "Execution.lbs{}".format(c_name)
+
+        try:
+            module = importlib.import_module(m_name)
+        except:
+            print("** ERROR: could not load module `{}`".format(
+                m_name))
+            return None
+
+        # Try to get concrete criterion class from module
+        try:
+            c_class = getattr(module, c_name)
+        except:
+            print("** ERROR: could not get class `{}` from module `{}`".format(
+                c_name,
+                m_name))
+            return None
+
+        # Instantiate and return object
+        ret_object = c_class(processors, edges, parameters)
+        print("[Criterion] Instantiated {} load transfer criterion".format(
+            c_name))
+        return ret_object
 
     ####################################################################
     @abc.abstractmethod
